refactor(db): Change DatabaseTrait methods to use &self
This moves the responsibility of synchronization from the caller to the
implementor.

The `commit` method on `DatabaseTrait` now takes `&self` instead of
`&mut self`. This allows thread-safe backends like RocksDB to be
used without an unnecessary, external Mutex wrapper.

The non-thread-safe `InMemoryDb` has been updated to use an internal
Mutex to manage its state, making it compliant with the new trait
bound.